yo yo

+++ src/main/java/com/example/management/StaffRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StaffRepository extends CrudRepository<Staff, Long> {
+public interface StaffRepository extends CrudRepository<Staff, Integer> {
 
 
 }
